requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic>=2.11.7,<3.0.0
pydantic-settings==2.1.0
sqlalchemy==2.0.25
python-dotenv==1.0.0
supabase==2.25.0
psycopg2-binary==2.9.9
elevenlabs==1.4.0
pyjwt>=2.10.1
google-generativeai
cryptography==42.0.5
mutagen==1.47.0

